making action tasks perform all their functions
